# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a API de gerenciamento de livros da Biblioteca.",
        "title": "API da Biblioteca",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/users/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON UserRequest e atualiza o usuario no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Atualiza um usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id da usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do novo usuario a ser atualizado",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario atualizado com sucesso"
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Exclui um usuario específico do banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Exclui um usuario pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuario a ser excluído",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nenhum Conteúdo"
                    },
                    "400": {
                        "description": "Requisição Inválida (ID com formato incorreto)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
...
        "/public/api/users": {
            "get": {
                "description": "Retorna uma lista de usuarios",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Listar usuarios",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/users.UserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON UserRequest e salva o usuario no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Cria um novo usuario",
                "parameters": [
                    {
                        "description": "Dados do Novo usuario a ser criado",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/users.UserRequest"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/public/api/users/login": {
            "post": {
                "description": "Recebe um objeto JSON LoginRequest.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Faz o login do usuario",
                "parameters": [
                    {
                        "description": "Dados para realizar o login",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token do usuario"
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/public/api/users/{id}": {
            "get": {
                "description": "Retorna um usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obter usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id do usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "middleware.APIError": {
            "description": "Erro padronizado de resposta.",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "users.LoginRequest": {
            "description": "Dados necessários para fazer o login",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "joaquim@email.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "users.Roles": {
            "type": "string",
            "enum": [
                "user",
                "admin"
            ],
            "x-enum-varnames": [
                "User",
                "Admin"
            ]
        },
        "users.UserRequest": {
            "description": "Dados necessários para criar usuario",
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "username"
            ],
            "properties": {
                "bio": {
                    "type": "string",
                    "example": "Meu nome é Joaquim"
                },
                "email": {
                    "type": "string",
                    "example": "joaquim@email.com"
                },
                "name": {
                    "type": "string",
                    "example": "Joaquim Silva"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "role": {
                    "$ref": "#/definitions/users.Roles"
                },
                "username": {
                    "type": "string",
                    "example": "joaoquim324"
                }
            }
        },
        "users.UserResponse": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/users.Roles"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}